# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Swapl API
  version: 0.1.0
  description: >-
    Shared contract for the Swapl home-swap API consumed by the web (Next.js),
    iOS, and Android clients. Server-authoritative: every platform calls this
    one API, which owns the single Postgres database. Schemas mirror the
    client DTOs proven to decode the live responses today.
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://swapl.fun
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque token from POST /api/auth/token (native clients).
    cookieAuth:
      type: apiKey
      in: cookie
      name: swapl_session
      description: Signed session cookie (web).
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
...
    PlanLimitError:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Machine-readable reason (e.g. PLAN_LIMIT). }
        upgradeTo: { type: string, description: Plan id that unlocks the feature. }
        currentPlan: { type: string }
    SavedSearch:
      type: object
      required: [id, userId, name, query, alertEnabled, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        name: { type: string }
        query: { type: string, description: Serialized URLSearchParams from /listings. }
        alertEnabled: { type: boolean }
        lastNotifiedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    SavedSearchesResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/SavedSearch' }
    SavedSearchCreateRequest:
      type: object
      required: [name, query]
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        query: { type: string, maxLength: 2000, description: Serialized URLSearchParams from /listings. }
        alertEnabled: { type: boolean, default: true }
    SavedSearchCreateResponse:
      type: object
      required: [ok, id]
      properties:
        ok: { type: boolean }
        id: { type: string }
...
paths:
...
  /api/saved-searches:
    get:
      tags: [saved-searches]
      summary: The user's saved searches
      description: Plus-plan feature; newest first.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SavedSearchesResponse' }
        '401': { description: Unauthenticated }
        '402':
          description: Plan does not include saved searches
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanLimitError' }
    post:
      tags: [saved-searches]
      summary: Save a search
      description: Plus-plan feature; capped at 20 per user.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SavedSearchCreateRequest' }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SavedSearchCreateResponse' }
        '400':
          description: Invalid input
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { description: Unauthenticated }
        '402':
          description: Plan does not include saved searches
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanLimitError' }
        '409':
          description: Saved-search cap reached (20)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/uploads/listing-photo:
    post:
      tags: [uploads]
      summary: Upload one listing photo (native clients)
      description: Single multipart image (max 8 MB) stored via UploadThing; the returned URL goes into listing.photos.
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Stored
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url: { type: string }
        '400': { description: Missing file }
        '401': { description: Unauthenticated }
        '413': { description: Image too large (max 8MB) }
        '415': { description: Not an image }
        '502': { description: Upload failed }
        '503': { description: Uploads not configured }
...